requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
psycopg2-binary
pandas
openpyxl==3.1.2
pytz==2023.3
python-dotenv==1.0.0
cryptography==41.0.7
Werkzeug==3.0.1